add explicit hint to not forget to uncomment the appId
The necessity of the appId is explanined in the docs linking to this page. But if you just come here from elsewhere this possible blocker can be overseen and can lead to unnecessary support requests, because it fails silently to display the dropdown even if it works with the default test data.